refactor simulation code into strength_analysis
